+++ test/test_qapi.py
@@ -1,19 +1,20 @@
 import os, unittest
-from qapi import QuranApi
+from source.qapi import QuranApi
+from source.qconfig import QConfig
 
 class TestQuranApi(unittest.TestCase):
     """Used to test the QuranApi class.
     """
 
-    def test_quran_api(self):
+    def test_quran_api(self) -> None:
         """Used to test all the methods of the QuranApi class
         """
 
-        # The absolute path to the database
-        cur_dir = os.path.dirname(os.path.realpath(__file__))
-        db_path = os.path.abspath(cur_dir + "/../data/quran.db")
+        # The application configuration
+        qconfig  = QConfig()
+        config   = qconfig.get_config()    
         # An instance of the QuranApi class is created
-        qapi = QuranApi(db_path, "ur")
+        qapi = QuranApi(config["db_path"], config["default_lang"])
 
         # The get_random_ruku method is tested 
         ruku_details = qapi.get_random_ruku()
@@ -47,12 +48,28 @@ class TestQuranApi(unittest.TestCase):
         self.assertEqual(len(sura_list), 114)
 
         # The set_lang method is tested 
-        qapi.set_lang("en")
+        qapi.set_lang("English")
         # Check that the tbl attribute has the correct value
         self.assertEqual(qapi.tbl, "ic_quranic_text-en")
         # Check that the lang attribute has the correct value
-        self.assertEqual(qapi.lang, "en")
-        
+        self.assertEqual(qapi.lang, "English")
+
+        # The get_lang_list method is tested 
+        lang_list = qapi.get_lang_list()
+        # Check that the number of languages is correct
+        self.assertEqual(len(lang_list), 43)
+
+        # The is_rtl method is tested 
+        is_rtl = qapi.is_rtl("Arabic")
+        # Check that the rtl vlaue is correct
+        self.assertTrue(is_rtl)
+
+        # The get_font_details method is tested 
+        font_details = qapi.get_font_details("Arabic")
+        # Check that the font family value is correct
+        self.assertEqual(font_details["family"], "Nafees [PYRS]")
+        # Check that the font size is correct
+        self.assertEqual(font_details["size"], 18)        
 
 if __name__ == '__main__':
     unittest.main()