+++ CHANGELOG.md
@@ -1,6 +1,22 @@
+# Islam Companion Desktop Reader 1.2.0
+
+## New features
+  
+  * Add all 43 languages to the language menu.
+  * Create config files for quran and hadith readers.
+  * Update the quran reader so it reads the font and table info from db.
+  * Update the quran reader so it reads the font files from a folder.
+
 # Islam Companion Desktop Reader 1.1.0
 
-## New Features
+## Minor bug fixes and improvements
+
+  * Update the path of the db file in qmanager.py and hmanager.py.
+  * Update the import statements so the code can be deployed.
+
+# Islam Companion Desktop Reader 1.1.0
+
+## New features
 
   * The quran and hadith readers need to support English, Arabic and Urdu languages.
   * The pyqtdeploy script does not seem to support sqlite3 Python module. The quran and hadith apis need to be updated so they use QtSql module of PyQt5.
@@ -10,4 +26,4 @@
 
 # Islam Companion Desktop Reader 1.0.0
 
-  * Initial Release.
+  * Initial release.